✓ Fixed primary sound register (SPB)
git-svn-id: http://amethyst.yui.ne.jp/svn/pc98/np2/branches/yui@1434 7c2116fe-53d0-4aec-ab59-80607f05abc0

+++ WORK_01/sound/fmboard.h
@@ -13,7 +13,7 @@
 
 
 typedef struct {
-	UINT16	addr1l;
+	UINT8	addr1l;
 	UINT8	addr1h;
 	UINT8	data1;
 	UINT8	addr1ext;
